+++ Packages/ai.muna.muna/Runtime/Beta/OpenAI/Types.cs
@@ -35,7 +35,6 @@ namespace Muna.Beta.OpenAI {
         /// <summary>
         /// Convert the `BinaryData` to a byte array.
         /// </summary>
-        /// <returns></returns>
         public byte[] ToArray() => data;
 
         /// <summary>