+++ Packages/ai.muna.muna/Runtime/Beta/Remote/RemotePredictionService.cs
@@ -69,52 +69,52 @@ namespace Muna.Beta.Services {
 
         internal RemotePredictionService(MunaClient client) => this.client = client;
 
-        private async Task<Value> ToValue(object? value) => value switch { // INCOMPLETE // Image
-            null              => new Value { type = Dtype.Null },
-            float           x => new Value { data = await Upload(new [] { x }.ToStream()), type = Dtype.Float32, shape = new int[0] },
-            double          x => new Value { data = await Upload(new [] { x }.ToStream()), type = Dtype.Float64, shape = new int[0] },
-            sbyte           x => new Value { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int8, shape = new int[0] },
-            short           x => new Value { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int16, shape = new int[0] },
-            int             x => new Value { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int32, shape = new int[0] },
-            long            x => new Value { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int64, shape = new int[0] },
-            byte            x => new Value { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint8, shape = new int[0] },
-            ushort          x => new Value { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint16, shape = new int[0] },
-            uint            x => new Value { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint32, shape = new int[0] },
-            ulong           x => new Value { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint64, shape = new int[0] },
-            bool            x => new Value { data = await Upload(new [] { x }.ToStream()), type = Dtype.Bool, shape = new int[0] },
-            float[]         x => new Value { data = await Upload(x.ToStream()), type = Dtype.Float32, shape = new [] { x.Length } },
-            double[]        x => new Value { data = await Upload(x.ToStream()), type = Dtype.Float64, shape = new [] { x.Length } },
-            sbyte[]         x => new Value { data = await Upload(x.ToStream()), type = Dtype.Int8, shape = new [] { x.Length } },
-            short[]         x => new Value { data = await Upload(x.ToStream()), type = Dtype.Int16, shape = new [] { x.Length } },
-            int[]           x => new Value { data = await Upload(x.ToStream()), type = Dtype.Int32, shape = new [] { x.Length } },
-            long[]          x => new Value { data = await Upload(x.ToStream()), type = Dtype.Int64, shape = new [] { x.Length } },
-            byte[]          x => new Value { data = await Upload(x.ToStream()), type = Dtype.Uint8, shape = new [] { x.Length } },
-            ushort[]        x => new Value { data = await Upload(x.ToStream()), type = Dtype.Uint16, shape = new [] { x.Length } },
-            uint[]          x => new Value { data = await Upload(x.ToStream()), type = Dtype.Uint32, shape = new [] { x.Length } },
-            ulong[]         x => new Value { data = await Upload(x.ToStream()), type = Dtype.Uint64, shape = new [] { x.Length } },
-            bool[]          x => new Value { data = await Upload(x.ToStream()), type = Dtype.Bool, shape = new [] { x.Length } },
-            Tensor<float>   x => new Value { data = await Upload(x.data.ToStream()), type = Dtype.Float32, shape = x.shape },
-            Tensor<double>  x => new Value { data = await Upload(x.data.ToStream()), type = Dtype.Float64, shape = x.shape },
-            Tensor<sbyte>   x => new Value { data = await Upload(x.data.ToStream()), type = Dtype.Int8, shape = x.shape },
-            Tensor<short>   x => new Value { data = await Upload(x.data.ToStream()), type = Dtype.Int16, shape = x.shape },
-            Tensor<int>     x => new Value { data = await Upload(x.data.ToStream()), type = Dtype.Int32, shape = x.shape },
-            Tensor<long>    x => new Value { data = await Upload(x.data.ToStream()), type = Dtype.Int64, shape = x.shape },
-            Tensor<byte>    x => new Value { data = await Upload(x.data.ToStream()), type = Dtype.Uint8, shape = x.shape },
-            Tensor<ushort>  x => new Value { data = await Upload(x.data.ToStream()), type = Dtype.Uint16, shape = x.shape },
-            Tensor<uint>    x => new Value { data = await Upload(x.data.ToStream()), type = Dtype.Uint32, shape = x.shape },
-            Tensor<ulong>   x => new Value { data = await Upload(x.data.ToStream()), type = Dtype.Uint64, shape = x.shape },
-            Tensor<bool>    x => new Value { data = await Upload(x.data.ToStream()), type = Dtype.Bool, shape = x.shape },
-            string          x => new Value { data = await Upload(x.ToStream(), mime: @"text/plain"), type = Dtype.String },
-            IList           x => new Value { data = await Upload(JsonConvert.SerializeObject(x).ToStream(), mime: @"application/json"), type = Dtype.List },
-            IDictionary     x => new Value { data = await Upload(JsonConvert.SerializeObject(x).ToStream(), mime: @"application/json"), type = Dtype.Dict },
-            Image           x => new Value { data = "", type = Dtype.Image },
-            Stream          x => new Value { data = await Upload(x), type = Dtype.Binary },
+        private async Task<RemoteValue> ToValue(object? value) => value switch { // INCOMPLETE // Image
+            null              => new() { type = Dtype.Null },
+            float           x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Float32, shape = new int[0] },
+            double          x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Float64, shape = new int[0] },
+            sbyte           x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int8, shape = new int[0] },
+            short           x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int16, shape = new int[0] },
+            int             x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int32, shape = new int[0] },
+            long            x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int64, shape = new int[0] },
+            byte            x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint8, shape = new int[0] },
+            ushort          x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint16, shape = new int[0] },
+            uint            x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint32, shape = new int[0] },
+            ulong           x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint64, shape = new int[0] },
+            bool            x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Bool, shape = new int[0] },
+            float[]         x => new() { data = await Upload(x.ToStream()), type = Dtype.Float32, shape = new [] { x.Length } },
+            double[]        x => new() { data = await Upload(x.ToStream()), type = Dtype.Float64, shape = new [] { x.Length } },
+            sbyte[]         x => new() { data = await Upload(x.ToStream()), type = Dtype.Int8, shape = new [] { x.Length } },
+            short[]         x => new() { data = await Upload(x.ToStream()), type = Dtype.Int16, shape = new [] { x.Length } },
+            int[]           x => new() { data = await Upload(x.ToStream()), type = Dtype.Int32, shape = new [] { x.Length } },
+            long[]          x => new() { data = await Upload(x.ToStream()), type = Dtype.Int64, shape = new [] { x.Length } },
+            byte[]          x => new() { data = await Upload(x.ToStream()), type = Dtype.Uint8, shape = new [] { x.Length } },
+            ushort[]        x => new() { data = await Upload(x.ToStream()), type = Dtype.Uint16, shape = new [] { x.Length } },
+            uint[]          x => new() { data = await Upload(x.ToStream()), type = Dtype.Uint32, shape = new [] { x.Length } },
+            ulong[]         x => new() { data = await Upload(x.ToStream()), type = Dtype.Uint64, shape = new [] { x.Length } },
+            bool[]          x => new() { data = await Upload(x.ToStream()), type = Dtype.Bool, shape = new [] { x.Length } },
+            Tensor<float>   x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Float32, shape = x.shape },
+            Tensor<double>  x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Float64, shape = x.shape },
+            Tensor<sbyte>   x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Int8, shape = x.shape },
+            Tensor<short>   x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Int16, shape = x.shape },
+            Tensor<int>     x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Int32, shape = x.shape },
+            Tensor<long>    x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Int64, shape = x.shape },
+            Tensor<byte>    x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Uint8, shape = x.shape },
+            Tensor<ushort>  x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Uint16, shape = x.shape },
+            Tensor<uint>    x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Uint32, shape = x.shape },
+            Tensor<ulong>   x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Uint64, shape = x.shape },
+            Tensor<bool>    x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Bool, shape = x.shape },
+            string          x => new() { data = await Upload(x.ToStream(), mime: @"text/plain"), type = Dtype.String },
+            IList           x => new() { data = await Upload(JsonConvert.SerializeObject(x).ToStream(), mime: @"application/json"), type = Dtype.List },
+            IDictionary     x => new() { data = await Upload(JsonConvert.SerializeObject(x).ToStream(), mime: @"application/json"), type = Dtype.Dict },
+            Image           x => new() { data = "", type = Dtype.Image },
+            Stream          x => new() { data = await Upload(x), type = Dtype.Binary },
             Enum            x => await ToValue(x.ToObject()),
-            Value           x => x,
+            RemoteValue     x => x,
             _                 => throw new InvalidOperationException($"Failed to serialize value '{value}' of type `{value.GetType()}` because it is not supported"),
         };
 
-        private async Task<object?> ToObject(Value value) { // INCOMPLETE // Image
+        private async Task<object?> ToObject(RemoteValue value) { // INCOMPLETE // Image
             if (value.type == Dtype.Null)
                 return null;
             using var stream = await Download(value.data!);
@@ -160,7 +160,7 @@ namespace Muna.Beta.Services {
 
         [Preserve, Serializable]
         private class RemotePrediction : Prediction {
-            public new Value[]? results;
+            public new RemoteValue[]? results;
         }
         #endregion
     }